% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{class}
\name{hetmetaObject}
\alias{hetmetaObject}
\title{hetmeta Object}
\description{
An object returned by \code{hetmeta} function, inheriting from class "\code{rma.uni}"
}
\details{
An object of class "\code{hetmeta}". The object is derived from an object of class \code{\link{rma.uni}}.
In addition to thatm it has the following components:
Objects of class "\code{hetmeta}" are lists with defined components.
\tabular{ll}{
\code{Rb} \tab value of \eqn{R_b}{Rb}, which quantifies the proportion of the between-study heterogeneity
relative to the variance of the pooled random effects estimate. \cr
\code{Ri} \tab value of \eqn{R_I}, whihc quantifies the proportion of the variance of the effect
 estimate due to between-studies variation. \cr
\code{CVb} \tab value of \eqn{CV_b}, the between-studies coefficient of variation.  \cr
\code{se_Rb} \tab the sandard error of \eqn{R_b2} derived using the delta method. \cr
\code{se_I2} \tab the sandard error of \eqn{I^2} derived using the delta method. \cr
\code{se_Ri} \tab the sandard error of \eqn{R_I} derived using the delta method. \cr
\code{se_CVb} \tab the sandard error of \eqn{CV_b} derived using the delta method. \cr
\code{s2_I2} \tab the "typical" within-study variance as defined in the \eqn{I^2} \cr
\code{s2_Ri} \tab the "typical" within-study variance as defined in the \eqn{R_I} \cr
\code{cv_vi} \tab value of the coefficient of variation of the within-study variances.  \cr
}
}
\author{
Alessio Crippa, \email{alessio.crippa@ki.se}
}
\seealso{
\code{\link{hetmeta}}, \code{\link{hetmeta-package}}
}

