% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.hetmeta}
\alias{print.hetmeta}
\title{Printing hetmeta Results}
\usage{
\method{print}{hetmeta}(x, digits, ...)
}
\arguments{
\item{x}{an object of class \code{hetmeta} produced by \code{\link{hetmeta}}.}

\item{digits}{an integer specifying the number of digits to which printed results must be rounded.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Print function for objects of class "\code{hetmeta}".
}
\examples{
#To be included
}
\author{
Alessio Crippa, \email{alessio.crippa@ki.se}
}
\seealso{
\code{\link{hetmeta}}
}

